feat: use proxy or message flow edit dialog depending on api type
