+++ ui-particles-angular/projects/ui-policy-studio-angular/src/lib/components/flow-form-dialog/flow-message-form-dialog/gio-ps-flow-message-form-dialog.component.ts
@@ -18,54 +18,56 @@ import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { FormControl, FormGroup } from '@angular/forms';
 import { cloneDeep, uniqueId } from 'lodash';
 
-import { FlowVM } from '../../gio-policy-studio.model';
-import { ConditionSelector, HttpMethod, HttpMethods, HttpSelector } from '../../models';
+import { FlowVM } from '../../../gio-policy-studio.model';
+import { ChannelSelector, ConditionSelector } from '../../../models';
+import { GioPolicyStudioFlowFormDialogResult } from '../gio-ps-flow-form-dialog-result.model';
 
-export type GioPolicyStudioFlowProxyFormDialogData = {
+export type GioPolicyStudioFlowMessageFormDialogData = {
   flow?: FlowVM;
+  entrypoints?: string[];
 };
 
-export type GioPolicyStudioFlowProxyFormDialogResult = FlowVM | false;
-
-type HttpMethodVM = HttpMethod | 'ALL';
-
 @Component({
-  selector: 'gio-ps-flow-proxy-form-dialog',
-  templateUrl: './gio-ps-flow-proxy-form-dialog.component.html',
-  styleUrls: ['./gio-ps-flow-proxy-form-dialog.component.scss'],
+  selector: 'gio-ps-flow-message-form-dialog',
+  templateUrl: './gio-ps-flow-message-form-dialog.component.html',
+  styleUrls: ['./gio-ps-flow-message-form-dialog.component.scss'],
 })
-export class GioPolicyStudioFlowProxyFormDialogComponent {
+export class GioPolicyStudioFlowMessageFormDialogComponent {
+  public entrypoints: string[] = [];
   public flowFormGroup?: FormGroup;
 
   public existingFlow?: FlowVM;
   public mode: 'create' | 'edit' = 'create';
-  public methods: HttpMethodVM[] = ['ALL', ...HttpMethods];
 
   constructor(
-    public dialogRef: MatDialogRef<GioPolicyStudioFlowProxyFormDialogComponent, GioPolicyStudioFlowProxyFormDialogResult>,
-    @Inject(MAT_DIALOG_DATA) flowDialogData: GioPolicyStudioFlowProxyFormDialogData,
+    public dialogRef: MatDialogRef<GioPolicyStudioFlowMessageFormDialogComponent, GioPolicyStudioFlowFormDialogResult>,
+    @Inject(MAT_DIALOG_DATA) flowDialogData: GioPolicyStudioFlowMessageFormDialogData,
   ) {
+    this.entrypoints = flowDialogData?.entrypoints ?? [];
+
     this.existingFlow = cloneDeep(flowDialogData?.flow);
     this.mode = this.existingFlow ? 'edit' : 'create';
 
-    const httpSelector = flowDialogData?.flow?.selectors?.find(s => s.type === 'HTTP') as HttpSelector;
+    const channelSelector = flowDialogData?.flow?.selectors?.find(s => s.type === 'CHANNEL') as ChannelSelector;
     const conditionSelector = flowDialogData?.flow?.selectors?.find(s => s.type === 'CONDITION') as ConditionSelector;
 
     this.flowFormGroup = new FormGroup({
       name: new FormControl(flowDialogData?.flow?.name ?? ''),
-      pathOperator: new FormControl(httpSelector?.pathOperator ?? 'EQUALS'),
-      path: new FormControl(httpSelector?.path ?? ''),
-      methods: new FormControl(httpSelector?.methods ?? ['ALL']),
+      channelOperator: new FormControl(channelSelector?.channelOperator ?? 'EQUALS'),
+      channel: new FormControl(channelSelector?.channel ?? ''),
+      operations: new FormControl(channelSelector?.operations ?? []),
+      entrypoints: new FormControl(channelSelector?.entrypoints ?? []),
       condition: new FormControl(conditionSelector?.condition ?? ''),
     });
   }
 
   public onSubmit(): void {
-    const httpSelectorToSave: HttpSelector = {
-      type: 'HTTP',
-      path: this.flowFormGroup?.get('path')?.value,
-      pathOperator: this.flowFormGroup?.get('pathOperator')?.value,
-      methods: sanitizeMethods(this.flowFormGroup?.get('methods')?.value),
+    const chanelSelectorToSave: ChannelSelector = {
+      type: 'CHANNEL',
+      channel: this.flowFormGroup?.get('channel')?.value,
+      channelOperator: this.flowFormGroup?.get('channelOperator')?.value,
+      operations: this.flowFormGroup?.get('operations')?.value,
+      entrypoints: this.flowFormGroup?.get('entrypoints')?.value,
     };
 
     const conditionValue: string = this.flowFormGroup?.get('condition')?.value;
@@ -85,7 +87,7 @@ export class GioPolicyStudioFlowProxyFormDialogComponent {
       _hasChanged: true,
       // Add changes
       name: this.flowFormGroup?.get('name')?.value,
-      selectors: conditionSelectorToSave ? [httpSelectorToSave, conditionSelectorToSave] : [httpSelectorToSave],
+      selectors: conditionSelectorToSave ? [chanelSelectorToSave, conditionSelectorToSave] : [chanelSelectorToSave],
       enabled: true,
     };
 
@@ -93,13 +95,4 @@ export class GioPolicyStudioFlowProxyFormDialogComponent {
       ...flowToSave,
     });
   }
-
-  public tagValidationHook(tag: string, validationCb: (shouldAddTag: boolean) => void) {
-    validationCb(HttpMethods.map(m => `${m}`).includes(tag.toUpperCase()));
-  }
 }
-
-const sanitizeMethods: (value: HttpMethodVM[]) => HttpMethod[] = (value?: HttpMethodVM[]) => {
-  if (!value || value.find(m => m === 'ALL')) return [];
-  return value as HttpMethod[];
-};
