+++ ui-particles-angular/projects/ui-policy-studio-angular/src/lib/components/flow-form-dialog/flow-message-form-dialog/gio-ps-flow-message-form-dialog.stories.ts
@@ -20,31 +20,32 @@ import { Component, Input } from '@angular/core';
 import { tap } from 'rxjs/operators';
 import { action } from '@storybook/addon-actions';
 
-import { GioPolicyStudioModule } from '../../gio-policy-studio.module';
-import { FlowVM } from '../../gio-policy-studio.model';
-import { fakeChannelFlow } from '../../models/index-testing';
+import { GioPolicyStudioModule } from '../../../gio-policy-studio.module';
+import { FlowVM } from '../../../gio-policy-studio.model';
+import { fakeChannelFlow } from '../../../models/index-testing';
+import { GioPolicyStudioFlowFormDialogResult } from '../gio-ps-flow-form-dialog-result.model';
 
 import {
-  GioPolicyStudioFlowProxyFormDialogComponent,
-  GioPolicyStudioFlowProxyFormDialogData,
-  GioPolicyStudioFlowProxyFormDialogResult,
-} from './gio-ps-flow-proxy-form-dialog.component';
+  GioPolicyStudioFlowMessageFormDialogComponent,
+  GioPolicyStudioFlowMessageFormDialogData,
+} from './gio-ps-flow-message-form-dialog.component';
 
 @Component({
-  selector: 'gio-ps-flow-proxy-form-dialog-story',
+  selector: 'gio-ps-flow-message-form-dialog-story',
   template: `<button id="open-dialog" (click)="openDialog()">Open dialog</button>`,
 })
-class GioPolicyStudioFlowProxyFormDialogStoryComponent {
+class GioPolicyStudioFlowMessageFormDialogStoryComponent {
   @Input() public flow?: FlowVM = undefined;
   constructor(private readonly matDialog: MatDialog) {}
 
   public openDialog() {
     this.matDialog
-      .open<GioPolicyStudioFlowProxyFormDialogComponent, GioPolicyStudioFlowProxyFormDialogData, GioPolicyStudioFlowProxyFormDialogResult>(
-        GioPolicyStudioFlowProxyFormDialogComponent,
+      .open<GioPolicyStudioFlowMessageFormDialogComponent, GioPolicyStudioFlowMessageFormDialogData, GioPolicyStudioFlowFormDialogResult>(
+        GioPolicyStudioFlowMessageFormDialogComponent,
         {
           data: {
             flow: this.flow,
+            entrypoints: ['entrypoint1', 'entrypoint2'],
           },
           role: 'alertdialog',
           id: 'gioPsFlowFormDialog',
@@ -61,17 +62,17 @@ class GioPolicyStudioFlowProxyFormDialogStoryComponent {
 }
 
 export default {
-  title: 'Policy Studio / components / Flow proxy form dialog',
-  component: GioPolicyStudioFlowProxyFormDialogStoryComponent,
+  title: 'Policy Studio / components / Flow message form dialog',
+  component: GioPolicyStudioFlowMessageFormDialogStoryComponent,
   decorators: [
     moduleMetadata({
-      declarations: [GioPolicyStudioFlowProxyFormDialogStoryComponent],
+      declarations: [GioPolicyStudioFlowMessageFormDialogStoryComponent],
       imports: [GioPolicyStudioModule, MatDialogModule, NoopAnimationsModule],
     }),
   ],
   argTypes: {},
   render: args => ({
-    component: GioPolicyStudioFlowProxyFormDialogStoryComponent,
+    component: GioPolicyStudioFlowMessageFormDialogStoryComponent,
     props: { ...args },
   }),
   parameters: {
