+++ ui-particles-angular/projects/ui-policy-studio-angular/src/lib/components/flow-form-dialog/flow-proxy-form-dialog/gio-ps-flow-proxy-form-dialog.component.html
@@ -48,7 +48,7 @@
     <mat-form-field>
       <mat-label>Path</mat-label>
       <input matInput formControlName="path" />
-      <mat-hint>Publish or Subscribe channel. This is the path of your request.</mat-hint>
+      <mat-hint>Path</mat-hint>
     </mat-form-field>
   </div>
