+++ ui-particles-angular/projects/ui-policy-studio-angular/src/lib/gio-policy-studio.module.ts
@@ -24,11 +24,11 @@ import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { ReactiveFormsModule } from '@angular/forms';
 
-import { GioPolicyStudioFlowMessageFormDialogComponent } from './components/flow-message-form-dialog/gio-ps-flow-message-form-dialog.component';
+import { GioPolicyStudioFlowMessageFormDialogComponent } from './components/flow-form-dialog/flow-message-form-dialog/gio-ps-flow-message-form-dialog.component';
 import { GioPolicyStudioDetailsComponent } from './components/flow-details/gio-ps-flow-details.component';
 import { GioPolicyStudioFlowsMenuComponent } from './components/flows-menu/gio-ps-flows-menu.component';
 import { GioPolicyStudioComponent } from './gio-policy-studio.component';
-import { GioPolicyStudioFlowProxyFormDialogComponent } from './components/flow-proxy-form-dialog/gio-ps-flow-proxy-form-dialog.component';
+import { GioPolicyStudioFlowProxyFormDialogComponent } from './components/flow-form-dialog/flow-proxy-form-dialog/gio-ps-flow-proxy-form-dialog.component';
 
 @NgModule({
   declarations: [
